add player controller with add, get all, get by id endpoints

+++ server/src/main/java/server/database/LobbyRepository.java
@@ -6,5 +6,5 @@ import server.entities.Lobby;
 import javax.transaction.Transactional;
 
 @Transactional
-public interface LobbyRepository extends JpaRepository<Lobby, Integer> {
+public interface LobbyRepository extends JpaRepository<Lobby, Long> {
 }
