+++ server/src/main/java/server/database/PlayerRepository.java
@@ -6,5 +6,7 @@ import server.entities.Player;
 import javax.transaction.Transactional;
 
 @Transactional
-public interface PlayerRepository extends JpaRepository<Player, Integer> {
+public interface PlayerRepository extends JpaRepository<Player, Long> {
+
+    //Optional<Player> findPlayerById(long playerId)
 }
